movies-search-engine/requirements.txt
pandas==1.4.2
weaviate_client==3.5.1
